Fix: 서버측 제공 데이터구조 변경으로 인한 수정
