+++ app/(afterLogin)/estimate/@estimateUser/components/EstimateList.tsx
@@ -1,10 +1,6 @@
 "use client";
-import {
-  EstimateItemAtom,
-  EstimateItemT,
-  fetchInfiniteEstimateT,
-  getTrainerEstimates,
-} from "../layout";
+import { EstimateItemAtom, EstimateItem } from "../../types";
+import { fetchInfiniteEstimateT } from "../layout";
 import EstimateUserAtom from "./EstimateUserAtom";
 import { InfiniteData, useInfiniteQuery } from "@tanstack/react-query";
 import { useInView } from "react-intersection-observer";
@@ -12,16 +8,16 @@ import { useEffect } from "react";
 
 export default function EstimateList() {
   const { data, fetchNextPage, hasNextPage, isFetching } = useInfiniteQuery<
-    EstimateItemT,
+    EstimateItem,
     Object,
-    InfiniteData<EstimateItemT>,
+    InfiniteData<EstimateItem>,
     Array<string>,
     number
   >({
     queryKey: ["trainerEstimates"],
     queryFn: fetchInfiniteEstimateT,
     initialPageParam: 0,
-    getNextPageParam: (lastPage: EstimateItemT) => {
+    getNextPageParam: (lastPage: EstimateItem) => {
       if (lastPage?.data[lastPage.data.length - 1]?.estimateId) {
         return lastPage?.data[lastPage.data.length - 1].estimateId;
       } else {
@@ -40,7 +36,7 @@ export default function EstimateList() {
   }, [inView, isFetching, hasNextPage, fetchNextPage]);
   return (
     <div className=" flex flex-col">
-      {data?.pages.map((page: EstimateItemT, pageIndex) =>
+      {data?.pages.map((page: EstimateItem, pageIndex) =>
         page.data.map((innerpage: EstimateItemAtom, itemIndex) => (
           <EstimateUserAtom
             data={innerpage}
